v1.02
* Travis CI: Update FASTBuild to v1.01

* Extending #if parsing with boolean && and || (#733)

* Extending #if parsing with boolean && and ||

* Fix build on gcc/g++

* Updated documentation

* Extending #if parsing with boolean && and ||

* Fix build on gcc/g++

* Updated documentation

* Apply && and || operator precedence (&& before ||)

* Tokenizer now recognises two character operators && and ||, #if handler modified accordingly

* Remove tests from if_directive.bff, add them into TestBFFParsing.cpp

* Actually evaluate the first defined result during the || pass

* Kludge tests a bit until better understood

* Formatting and cleanup

* Added errors for #if operator parsing, applied them to tests.


* Fix bad merge

* Delete stale file [skip ci]

* Post intergation: #if supports || and &&

 - simplify logic
 - add tests for some additional cases
 - simplify tests
 - code style
 - add documentation for new errors
 - update documentation to reflect precedence rules

* [Improvement] Increase thread stack size to avoid unexpected problems with cache plugins

* CachePlugin improvements

[Note] DB Version changed
[Improvement] Cache plugin can be initialized via InitEx which provides details of operating mode (read/write/verbose)
[Improvement] Settings exposes .CachePluginDLLConfig to all passing of custom config to cache plugins
[Improvement][Windows] Cache plugin DLL function names no longer need to be mangled

* Fix tests for VS older that VS2019

 - __pragma in macro behavior seems inconsistent
 - just do things manually for now

* Replace master with better wording (#756)

* Replace master with better wording

* Force rebuild

* Replace blacklist with better wording (#757)

* Added “AllowResponseFile” setting. (#501)


* Post-integration cleanup: AllowResponseFiles

 - add documentation
 - fix unreachable code
 - update integraiton/markup files
 - increment database version

* Fix typo in test bff [skip ci]

* [Improvement] ObjectLists can be chained using .CompilerInputObjectLists

* Fix memory leak in cache test

* Code style cleanup:

 - remove use of auto
 - replace some manual loops with range-based for
 - add some missing const

* LZ4: Remove .gitignore and Makefile

* [Fix][Windows] Prevent deadlock in FBuildWorker on shutdown when it's sent a WM_QUIT by another application

* [Improvement] VCXProject and XCodeProject include .editorconfig files by default

[Fix] Fix documentation errors for VCXProject and XCodeProject

* Update Visual Studio 2019 to 14.26.28801 (16.6.4)

* Emit monitor events for EXEC_NODE (#768)

* [WIP] Add new `docs/troubleshooting/cache.html` file to describe `cache debugging` strategies (#729)

This document is a direct result of the discussions and debugging here:
https://github.com/fastbuild/fastbuild/issues/727

NOTE: There remain some `TODO`(s) and unfinished sections to this file.
I don't have all of the knowledge required to finish these sections
properly. Hopefully @ffulin can fill in the blanks here.

* Update `docs/documentation.html` to add a hyperlink to this new file.


* Further improvements to Cache troubleshooting documentation

* Documentation styling improvements and fixes

* Add .SourceMapping_Experimental to Compiler node (#769)

* Post integration: Cleanup SourceMapping_Experimental

 - minor code style fixes
 - enable test on Windows with Clang
 - disable distributed compilation temporartily when used and document that

* [Fix] ForEach correctly handles comma separator in parallel loops

* Update FASTBuild to use VS2019 16.7.0 (14.27.29110)

 - fix various newly surfaced compiler warnings

* Fix *ExcludePattern inconsistencies with slash direction

[Fix] CSAssembly() handles forward and backslashes correctly on all platforms in .CompilerInputExcludePattern
[Fix] Exec() handles forward and backslashes correctly on all platforms in .ExecInputExcludePattern
[Fix] ObjectList() handles forward and backslashes correctly on all platforms in .CompilerInputExcludePattern
[Fix] Test() handles forward and backslashes correctly on all platforms in .TestInputExcludePattern

* Generate integration files for recently added SourceMapping_Experimental [skip ci]

* Response file improvements

[Improvement] Compiler() response file use can be forced with .ForceResponseFile
[Improvement] Library() response file use can be forced with .LibrarianForceResponseFile
[Improvement] Executable()/DLL() response file use can be forced with .LinkerForceResponseFile
[Improvement][Linux] Command line limits (128 KiB) are enforced on Linux to make problems more obvious
[Fix][Linux] Various "CanUseResponseFile" options are respected on Linux

* Possible fix for file permissions problems on CI

Uniquify response file names over time as a possible work around for permissions isuses on some machines
 - files reliably fail to open on the CI for some reason when re-used in rapid succession
 - this may be related to security software of some kind

* Increase timeout on Response file permissions workaround for Windows

* Improve output on failing assertion

* Enhance assertion message further

* Fix Args test

 - prevent tests executing code that relies on WorkerThread having previously created directories
 - revert some changes made during troubleshooting this issue

* [Fix] Fix a failure path writing response files that was not stopping compilation correctly

* Update documentation for ForceResponseFile options [skip ci]

* v1.02

+++ Code/Core/Math/Random.h
@@ -31,7 +31,7 @@ public:
     // random index from 0 to size-1
     uint32_t GetRandIndex( uint32_t size )
     {
-        return ( (uint32_t)( (float)size * ( GetRand() / ( CORE_RAND_MAX + 1.0f ) ) ) );
+        return ( (uint32_t)( (float)size * ( (float)GetRand() / (float)( CORE_RAND_MAX + 1 ) ) ) );
     }
 
     // access the seed value
